# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ApplicationResponse": {
        "properties": {
          "allowed_origins": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Origins",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "public_key": {
            "title": "Public Key",
            "type": "string"
          },
          "rate_limit_per_minute": {
            "title": "Rate Limit Per Minute",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "name",
          "public_key",
          "allowed_origins",
          "rate_limit_per_minute",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ApplicationResponse",
        "type": "object"
      },
...
      "Body_uploadKnowledgeDocument": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "replace_document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replace Document Id"
          },
          "source_url": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "file",
          "title"
        ],
        "title": "Body_uploadKnowledgeDocument",
        "type": "object"
      },
...
      "DocumentAcceptedResponse": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentResponse"
          },
          "job": {
            "$ref": "#/components/schemas/IngestionJobResponse"
          }
        },
        "required": [
          "document",
          "job"
        ],
        "title": "DocumentAcceptedResponse",
        "type": "object"
      },
      "DocumentResponse": {
        "properties": {
          "byte_size": {
            "title": "Byte Size",
            "type": "integer"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "knowledge_base_id": {
            "format": "uuid",
            "title": "Knowledge Base Id",
            "type": "string"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "source_filename": {
            "title": "Source Filename",
            "type": "string"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "supersedes_document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Document Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "knowledge_base_id",
          "supersedes_document_id",
          "version",
          "title",
          "source_filename",
          "source_url",
          "mime_type",
          "byte_size",
          "content_hash",
          "status",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KnowledgeBaseCreate": {
        "properties": {
          "description": {
            "default": "",
            "maxLength": 4000,
            "title": "Description",
            "type": "string"
          },
          "embedding_model_config_id": {
            "format": "uuid",
            "title": "Embedding Model Config Id",
            "type": "string"
          },
          "embedding_version": {
            "default": "v1",
            "maxLength": 64,
            "minLength": 1,
            "title": "Embedding Version",
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "embedding_model_config_id"
        ],
        "title": "KnowledgeBaseCreate",
        "type": "object"
      },
      "KnowledgeBaseResponse": {
        "properties": {
          "chunking_version": {
            "title": "Chunking Version",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "embedding_dimension": {
            "title": "Embedding Dimension",
            "type": "integer"
          },
          "embedding_model_config_id": {
            "format": "uuid",
            "title": "Embedding Model Config Id",
            "type": "string"
          },
          "embedding_model_name": {
            "title": "Embedding Model Name",
            "type": "string"
          },
          "embedding_version": {
            "title": "Embedding Version",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/KnowledgeBaseStatus"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "name",
          "description",
          "embedding_model_config_id",
          "embedding_model_name",
          "embedding_dimension",
          "embedding_version",
          "chunking_version",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "KnowledgeBaseResponse",
        "type": "object"
      },
...
      "KnowledgeBaseUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeBaseStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "KnowledgeBaseUpdate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "AI Customer Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/knowledge-bases": {
      "get": {
        "operationId": "listKnowledgeBases",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeBaseResponse"
                  },
                  "title": "Response Listknowledgebases",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Knowledge Bases",
        "tags": [
          "knowledge-bases"
        ]
      },
      "post": {
        "operationId": "createKnowledgeBase",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Knowledge Base",
        "tags": [
          "knowledge-bases"
        ]
      }
    },
    "/v1/admin/knowledge-bases/{base_id}": {
      "patch": {
        "operationId": "updateKnowledgeBase",
        "parameters": [
          {
            "in": "path",
            "name": "base_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Base Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Knowledge Base",
        "tags": [
          "knowledge-bases"
        ]
      }
    },
    "/v1/admin/knowledge-bases/{base_id}/applications": {
      "get": {
        "operationId": "listKnowledgeBaseApplications",
        "parameters": [
          {
            "in": "path",
            "name": "base_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Base Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApplicationResponse"
                  },
                  "title": "Response Listknowledgebaseapplications",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Knowledge Base Applications",
        "tags": [
          "knowledge-bases"
        ]
      }
    },
    "/v1/admin/knowledge-bases/{base_id}/applications/{application_id}": {
      "delete": {
        "operationId": "unbindKnowledgeBase",
        "parameters": [
          {
            "in": "path",
            "name": "base_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Base Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Unbind Knowledge Base",
        "tags": [
          "knowledge-bases"
        ]
      },
      "put": {
        "operationId": "bindKnowledgeBase",
        "parameters": [
          {
            "in": "path",
            "name": "base_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Base Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Bind Knowledge Base",
        "tags": [
          "knowledge-bases"
        ]
      }
    },
    "/v1/admin/knowledge-bases/{base_id}/documents": {
      "get": {
        "operationId": "listKnowledgeDocuments",
        "parameters": [
          {
            "in": "path",
            "name": "base_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Base Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  },
                  "title": "Response Listknowledgedocuments",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Documents",
        "tags": [
          "knowledge-bases"
        ]
      },
      "post": {
        "operationId": "uploadKnowledgeDocument",
        "parameters": [
          {
            "in": "path",
            "name": "base_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Base Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadKnowledgeDocument"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Document",
        "tags": [
          "knowledge-bases"
        ]
      }
    },
    "/v1/admin/knowledge-bases/{base_id}/documents/{document_id}": {
      "delete": {
        "operationId": "deleteKnowledgeDocument",
        "parameters": [
          {
            "in": "path",
            "name": "base_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Base Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Document",
        "tags": [
          "knowledge-bases"
        ]
      },
      "get": {
        "operationId": "getKnowledgeDocument",
        "parameters": [
          {
            "in": "path",
            "name": "base_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Base Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Document",
        "tags": [
          "knowledge-bases"
        ]
      }
    },
    "/v1/admin/knowledge-bases/{base_id}/documents/{document_id}/retry": {
      "post": {
        "operationId": "retryKnowledgeDocumentIngestion",
        "parameters": [
          {
            "in": "path",
            "name": "base_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Base Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retry Document",
        "tags": [
          "knowledge-bases"
        ]
      }
    },
    "/v1/admin/knowledge-bases/{base_id}/search": {
      "post": {
        "operationId": "searchKnowledgeBase",
        "parameters": [
          {
            "in": "path",
            "name": "base_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Base Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SearchResultResponse"
                  },
                  "title": "Response Searchknowledgebase",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search Knowledge Base",
        "tags": [
          "knowledge-bases"
        ]
      }
    },
...
  }
}